Exceptions

+++ src/main/java/com/desafio/CRUD/clientes/service/ClientService.java
@@ -3,6 +3,8 @@ package com.desafio.CRUD.clientes.service;
 import com.desafio.CRUD.clientes.dto.ClientDTO;
 import com.desafio.CRUD.clientes.entities.Client;
 import com.desafio.CRUD.clientes.repositories.ClientRepository;
+import com.desafio.CRUD.clientes.service.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -18,9 +20,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        return repository.findById(id)
-                .map(ClientDTO::new)
-                .orElse(new ClientDTO()); // retorna DTO vazio se não achar
+        Client client = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
+        return new ClientDTO(client);
     }
 
     @Transactional(readOnly = true)
@@ -39,14 +40,26 @@ public class ClientService {
 
     @Transactional
     public ClientDTO update(Long id, ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ClientDTO(entity);
+
+        try {
+
+            Client entity = repository.getReferenceById(id);
+            copyDtoEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ClientDTO(entity);
+
+        }
+        catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontado");
+        }
+
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         repository.deleteById(id);
     }
